fix artifact output

+++ huff_core/src/lib.rs
@@ -371,16 +371,32 @@ impl<'a> Compiler {
     /// 1. Cleans any previous artifacts in the output directory.
     /// 2. Exports artifacts in parallel as serialized json `Artifact` objects.
     pub fn export_artifacts(artifacts: &Vec<Arc<Artifact>>, output: &OutputLocation) {
+        // Exit if empty output location
+        if output.0.is_empty() {
+            tracing::warn!(target: "core", "Exiting artifact export with empty output location!");
+            return
+        }
+
         // Clean the Output Directory
         tracing::warn!(target: "core", "REMOVING DIRECTORY: \"{}\"", output.0);
         if !output.0.is_empty() && fs::remove_dir_all(&output.0).is_ok() {
             tracing::info!(target: "core", "OUTPUT DIRECTORY DELETED!");
         }
 
+        // Is the output a directory or a file?
+        let is_file = std::path::PathBuf::from(&output.0).extension().is_some();
+
         // Export the artifacts with parallelized io
         artifacts.into_par_iter().for_each(|a| {
-            let json_out =
-                format!("{}/{}.json", output.0, a.file.path.to_uppercase().replacen("./", "", 1));
+            // If it's a file type, we just export to `output.0`
+            let json_out = match is_file {
+                true => output.0.clone(),
+                false => format!(
+                    "{}/{}.json",
+                    output.0,
+                    a.file.path.to_uppercase().replacen("./", "", 1)
+                ),
+            };
             tracing::debug!(target: "core", "JSON OUTPUT: {:?}", json_out);
 
             if let Err(e) = a.export(&json_out) {
